Consolidate nightly monitoring into ci_analyzer.py
- Integrate nightly monitoring functionality into ci_analyzer.py with --mode flag
- Update nightly job names for Nvidia, AMD, and Intel workflows
- Remove standalone nightly_monitor.py file
- Update workflow files to use ci_analyzer.py --mode nightly
