+++ .github/workflows/test-nightly-monitor.yml
@@ -43,8 +43,9 @@ jobs:
           PYTHONIOENCODING: utf-8
         run: |
           cd scripts/ci_monitor
-          python nightly_monitor.py \
+          python ci_analyzer.py \
             --token $GITHUB_TOKEN \
+            --mode nightly \
             --days ${{ inputs.days || '7' }} \
             --output test_monitor_$(date +%Y%m%d_%H%M%S).json
